+++ activity-service/src/main/java/com/example/activityservice/vo/sleep/ResponseSleep.java
@@ -11,5 +11,5 @@ import java.time.LocalDateTime;
 public class ResponseSleep {
     private LocalDateTime sleptTime;
     private LocalDateTime wakeUpTime;
-    private int totalSleepTime;
+    private Long totalSleepTime;
 }
